updating distribute function.  cleaning up code.
