+++ Project.toml
@@ -1,6 +1,6 @@
-name = "PipsSolver"
+name = "PipsNLP"
 uuid = "0e2b999c-8a3a-11ea-3c4b-c36350f324a0"
-authors = ["Jordan Jalving", "University of Wisconsin-Madison"]
+authors = ["Jordan Jalving and Yankai Cao", "University of Wisconsin-Madison"]
 version = "0.1.0"
 
 [deps]
@@ -16,7 +16,12 @@ Plasmo = "d3f7391f-f14a-50cc-bbe4-76a32d1bad3c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
-Plasmo = "0.3.0,0.4.0"
+DataStructures = "0.18"
+JuMP = "0.19, 0.20, 0.21"
+MPI = "0.14"
+MPIClusterManagers = "0.2"
+MathOptInterface = "0.9"
+Plasmo = "0.3, 0.4"
 julia = "1"
 
 [extras]
