+++ src/PipsNlpInterface.jl
@@ -1,6 +1,7 @@
 module PipsNlpInterface
 
-#Import Packages
+#TODO: update pips-nlp interface to work on optigraphs in addition to optinodes for subproblems
+
 using SparseArrays
 using LinearAlgebra
 using DataStructures
@@ -15,221 +16,166 @@ include("PipsNlpCInterface.jl")
 using .PipsNlpSolver
 
 include("pips_utils.jl")
+include("pips_nlp_data.jl")
 include("nlp_evaluator.jl")
 
 export pipsnlp_solve
-
-mutable struct PIPSNLPData
-    d::Union{JuMP.NLPEvaluator,Nothing}
-    n::Int
-    m::Int
-    local_m::Int                    #rank row dimension
-    jacnnz::Int                     #Jacobian num nonzeros
-    hessnnz::Int                    #Hessian num nonzeros
-    firstIeq::Vector{Int}           #first stage equality constraint jacobian nonzero rows
-    firstJeq::Vector{Int}           #first stage equality constraint jacobian nonzero cols
-    firstVeq::Vector{Float64}       #first stage equality constraint jacobian values
-    secondIeq::Vector{Int}          #second stage equality constraint jacobian nonzero rows
-    secondJeq::Vector{Int}          #second stage equality constraint jacobian nonzero cols
-    secondVeq::Vector{Float64}      #second stage equality constraint jacobian values
-    firstIineq::Vector{Int}         #first stage inequality constraint jacobian nonzero rows
-    firstJineq::Vector{Int}
-    firstVineq::Vector{Float64}
-    secondIineq::Vector{Int}
-    secondJineq::Vector{Int}
-    secondVineq::Vector{Float64}
-    num_eqconnect::Int              #num equalities connecting first to second stage
-    num_ineqconnect::Int            #num inequalities connecting first to second stage
-    eqconnect_lb::Vector{Float64}   #connection equality lower bounds
-    eqconnect_ub::Vector{Float64}   #connection equality upper bounds
-    ineqconnect_lb::Vector{Float64} #connection inequality lower bounds
-    ineqconnect_ub::Vector{Float64} #connection inequality upper bounds
-    eq_idx::Vector{Int}             #equality connect indices
-    ineq_idx::Vector{Int}           #inequality connect indices
-    firstJeqmat
-    secondJeqmat
-    firstJineqmat
-    secondJineqmat
-    linkIeq::Vector{Int}            #equality link jacobian rows
-    linkJeq::Vector{Int}            #equality link jacobian cols
-    linkVeq::Vector{Float64}        #equality link jacobian values
-    linkIineq::Vector{Int}
-    linkJineq::Vector{Int}
-    linkVineq::Vector{Float64}
-    x_sol::Vector{Float64}          #solution
-    coreid::Int                     #rank
-    loaded::Bool
-    local_unsym_hessnnz::Int        #local hessian num nonzeros
-end
-PIPSNLPData() = PIPSNLPData(nothing,0,0,0,0,0,Int[],Int[], Float64[], Int[], Int[], Float64[],Int[],Int[],Float64[], Int[], Int[],
-                            Float64[], 0, 0, Float64[],Float64[],Float64[],Float64[], Int[], Int[],
-                            nothing, nothing, nothing, nothing, Int[],Int[], Float64[], Int[], Int[], Float64[],Float64[], 0, false, 0)
-
-
-function _get_pips_data(node::OptiNode)
-    return node.ext[:pips_data]
-end
 
 function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints are first stage
     #Check structure: 3 possibilities
     #1.) optigraph without subgraphs (only link constraints, no first stage -> second stage constraints)
     #2.) optigraph with single subgraph (first stage and linkconstraints between optinodes)
-    #3.) optigraph with multiple subgraphs (first stage and linkconstraints between subgraphs) Not yet supported
-
-    if !(has_subgraphs(graph))
-        first_stage = OptiNode() #empty first stage
-        submodels = all_nodes(graph)
-    elseif length(graph.subgraphs) == 1
-        num_nodes(graph) == 1 || error("Incompatible optigraph structure.  Optigraph contains a single subgraph, but has multiple optinodes in highest level graph.")
-        first_stage = getnode(graph,1)
-        submodels = all_nodes(graph.subgraphs[1])
-    end
+    #3.) optigraph with multiple subgraphs (first stage and linkconstraints between subgraphs). Not yet supported
 
     comm = MPI.COMM_WORLD
     if MPI.Comm_rank(comm) == 0
         println("Building Model for PIPS-NLP")
     end
 
-    n_sub_models = length(submodels)
+    if haskey(graph.ext[:user_pips_data])
+        pips_data = graph.ext[:user_pips_data]
+    else
+        pips_data = _setup_pips_nlp_data(graph)
+    end
 
+    first_stage = pips_data.first_stage
+    sub_models = pips_data.submodels
+    n_sub_models = length(submodels)
     model_list = [first_stage; submodels]
 
-    for (idx,node) in enumerate(model_list)
-        node.ext[:pips_data] = PIPSNLPData()
-        constraint_data = get_constraint_data(node)
-        node.ext[:constraint_data] = constraint_data
-    end
     first_stage_data = _get_pips_data(first_stage)
+    eqlink_lb = pips_data.
 
-    #Assume every worker has every linkconstraint that connects across nodes
-    #TODO: MPI.ALLGATHER! to setup linkconstraints when they are different across workers
-    #TODO: Check whether workers can use zeros for bounds they don't need.
-    #TODO: Improve distribute performance.
+    #We assume every worker has the link constraint structure that connects across nodes
+    #TODO: use MPI.ALLGATHER! to setup linkconstraints when they are different across workers?
+    #NOTE: Each rank would need to know which links are equivalent e.g. using some sort of user-defined index.  For now,
+    #we assume the user needs to model their problem correctly using ghost optinodes, or they use distribute_optigraph, which sets up
+    #links correctly.  In the future, distributed optigraphs will make communicating the structure easier.
+
+    #TODO: Check whether workers can use zeros for bounds on linkconstraints they don't need.
 
     #IDENTIFY LINK CONSTRAINTS
-    link_connect_eq,link_connect_ineq,linkeqconstraints,linkineqconstraints  = _identify_linkconstraints(graph)
+    # link_connect_eq,link_connect_ineq,linkeqconstraints,linkineqconstraints  = _identify_linkconstraints(graph)
+    #
+    # #FIRST STAGE -> SECOND STAGE LINK CONSTRAINTS
+    # for link in link_connect_eq
+    #     node = _get_subnode(first_stage,link)::OptiNode
+    #     local_data = _get_pips_data(node)
+    #     firstIeq = local_data.firstIeq
+    #     firstJeq = local_data.firstJeq
+    #     firstVeq = local_data.firstVeq
+    #     secondIeq = local_data.secondIeq
+    #     secondJeq = local_data.secondJeq
+    #     secondVeq = local_data.secondVeq
+    #     push!(local_data.eqconnect_lb, link.set.value)
+    #     push!(local_data.eqconnect_ub, link.set.value)
+    #     local_data.num_eqconnect += 1
+    #     row = local_data.num_eqconnect
+    #     for (var,coeff) in link.func.terms
+    #         if getnode(var) == first_stage
+    #             push!(firstIeq, row)
+    #             push!(firstJeq, var.index.value)
+    #             push!(firstVeq, coeff)
+    #         else
+    #             @assert getnode(var) == node
+    #             push!(secondIeq, row)
+    #             push!(secondJeq, var.index.value)
+    #             push!(secondVeq, coeff)
+    #         end
+    #     end
+    # end
+    #
+    # for link in link_connect_ineq
+    #     node = _get_subnode(first_stage,link)
+    #     firstIineq = local_data.firstIineq
+    #     firstJineq = local_data.firstJineq
+    #     firstVineq = local_data.firstVineq
+    #     secondIineq = local_data.secondIineq
+    #     secondJineq = local_data.secondJineq
+    #     secondVineq = local_data.secondVineq
+    #     if isa(link.set,MOI.LessThan)
+    #         push!(local_data.ineqconnect_lb, -Inf)
+    #         push!(local_data.ineqconnect_ub, link.set.upper)
+    #     elseif isa(link.set,MOI.GreaterThan)
+    #         push!(local_data.ineqconnect_lb, link.set.lower)
+    #         push!(local_data.ineqconnect_ub, Inf)
+    #     elseif isa(link.set,MOI.Interval)
+    #         push!(local_data.ineqconnect_lb, link.set.lower)
+    #         push!(local_data.ineqconnect_ub, link.set.upper)
+    #     end
+    #     local_data.num_ineqconnect += 1
+    #     row = local_data.num_ineqconnect
+    #     for (var,coeff) in link.func.terms
+    #         if getnode(var) == first_stage
+    #             push!(firstIineq, row)
+    #             push!(firstJineq, var.index.value)
+    #             push!(firstVineq, ind)
+    #         else
+    #             @assert getnode(var) == node
+    #             push!(secondIineq, row)
+    #             push!(secondJineq, var.index.value)
+    #             push!(secondVineq, ind)
+    #         end
+    #     end
+    # end
+    #
+    # #SECOND STAGE LINK CONSTRAINTS
+    # nlinkeq = length(linkeqconstraints)
+    # nlinkineq = length(linkineqconstraints)
+    # ineqlink_lb = zeros(nlinkineq)
+    # ineqlink_ub = zeros(nlinkineq)
+    # eqlink_lb = zeros(nlinkeq)
+    # eqlink_ub = zeros(nlinkeq)
+    #
+    # #INEQUALITY CONSTRAINTS
+    # #Inequality bounds
+    # #NOTE: linkconstraint bounds need to be on every rank.
+    # for (idx,link) in enumerate(linkineqconstraints)
+    #     row = idx
+    #     #Bounds
+    #     if isa(link.set,MOI.LessThan)
+    #         ineqlink_lb[row] = -Inf
+    #         ineqlink_ub[row] = link.set.upper
+    #     elseif isa(link.set,MOI.GreaterThan)
+    #         ineqlink_lb[row] = link.set.lower
+    #         ineqlink_ub[row] = Inf
+    #     elseif isa(link.set,MOI.Interval)
+    #         ineqlink_lb[row] = link.set.lower
+    #         ineqlink_ub[row] = link.set.upper
+    #     end
+    #     #Inequality values
+    #     for (var,coeff) in link.func.terms
+    #         node = getnode(var)
+    #         local_data = _get_pips_data(node)
+    #         linkIineq = local_data.linkIineq
+    #         linkJineq = local_data.linkJineq
+    #         linkVineq = local_data.linkVineq
+    #         push!(linkIineq, row)
+    #         push!(linkJineq, var.index.value)
+    #         push!(linkVineq, coeff)
+    #     end
+    # end
+    #
+    # #EQUALITY CONSTRAINTS
+    # for (idx,link) in enumerate(linkeqconstraints)
+    #     row = idx
+    #     #Bounds
+    #     eqlink_lb[row] = link.set.value
+    #     eqlink_ub[row] = link.set.value
+    #     #Equality values
+    #     for (var,coeff) in link.func.terms
+    #         node = getnode(var)
+    #         local_data = _get_pips_data(node)
+    #         linkIeq = local_data.linkIeq
+    #         linkJeq = local_data.linkJeq
+    #         linkVeq = local_data.linkVeq
+    #         push!(linkIeq, row)              #the variable row
+    #         push!(linkJeq, var.index.value)  #the variable column
+    #         push!(linkVeq, coeff)            #the coefficient
+    #     end
+    # end
 
-
-    #FIRST STAGE -> SECOND STAGE LINK CONSTRAINTS
-    for link in link_connect_eq
-        node = _get_subnode(first_stage,link)::OptiNode
-        local_data = _get_pips_data(node)
-        firstIeq = local_data.firstIeq
-        firstJeq = local_data.firstJeq
-        firstVeq = local_data.firstVeq
-        secondIeq = local_data.secondIeq
-        secondJeq = local_data.secondJeq
-        secondVeq = local_data.secondVeq
-        push!(local_data.eqconnect_lb, link.set.value)
-        push!(local_data.eqconnect_ub, link.set.value)
-        local_data.num_eqconnect += 1
-        row = local_data.num_eqconnect
-        for (var,coeff) in link.func.terms
-            if getnode(var) == first_stage
-                push!(firstIeq, row)
-                push!(firstJeq, var.index.value)
-                push!(firstVeq, coeff)
-            else
-                @assert getnode(var) == node
-                push!(secondIeq, row)
-                push!(secondJeq, var.index.value)
-                push!(secondVeq, coeff)
-            end
-        end
-    end
-
-    for link in link_connect_ineq
-        node = _get_subnode(first_stage,link)
-        firstIineq = local_data.firstIineq
-        firstJineq = local_data.firstJineq
-        firstVineq = local_data.firstVineq
-        secondIineq = local_data.secondIineq
-        secondJineq = local_data.secondJineq
-        secondVineq = local_data.secondVineq
-        if isa(link.set,MOI.LessThan)
-            push!(local_data.ineqconnect_lb, -Inf)
-            push!(local_data.ineqconnect_ub, link.set.upper)
-        elseif isa(link.set,MOI.GreaterThan)
-            push!(local_data.ineqconnect_lb, link.set.lower)
-            push!(local_data.ineqconnect_ub, Inf)
-        elseif isa(link.set,MOI.Interval)
-            push!(local_data.ineqconnect_lb, link.set.lower)
-            push!(local_data.ineqconnect_ub, link.set.upper)
-        end
-        local_data.num_ineqconnect += 1
-        row = local_data.num_ineqconnect
-        for (var,coeff) in link.func.terms
-            if getnode(var) == first_stage
-                push!(firstIineq, row)
-                push!(firstJineq, var.index.value)
-                push!(firstVineq, ind)
-            else
-                @assert getnode(var) == node
-                push!(secondIineq, row)
-                push!(secondJineq, var.index.value)
-                push!(secondVineq, ind)
-            end
-        end
-    end
-
-    #SECOND STAGE LINK CONSTRAINTS
-    nlinkeq = length(linkeqconstraints)
-    nlinkineq = length(linkineqconstraints)
-    ineqlink_lb = zeros(nlinkineq)
-    ineqlink_ub = zeros(nlinkineq)
-    eqlink_lb = zeros(nlinkeq)
-    eqlink_ub = zeros(nlinkeq)
-
-    #INEQUALITY CONSTRAINTS
-    #Inequality bounds
-    #NOTE: linkconstraint bounds need to be on every rank.
-    for (idx,link) in enumerate(linkineqconstraints)
-        row = idx
-        if isa(link.set,MOI.LessThan)
-            ineqlink_lb[row] = -Inf
-            ineqlink_ub[row] = link.set.upper
-        elseif isa(link.set,MOI.GreaterThan)
-            ineqlink_lb[row] = link.set.lower
-            ineqlink_ub[row] = Inf
-        elseif isa(link.set,MOI.Interval)
-            ineqlink_lb[row] = link.set.lower
-            ineqlink_ub[row] = link.set.upper
-        end
-        #Inequality values
-        for (var,coeff) in link.func.terms
-            node = getnode(var)
-            local_data = _get_pips_data(node)
-            linkIineq = local_data.linkIineq
-            linkJineq = local_data.linkJineq
-            linkVineq = local_data.linkVineq
-            push!(linkIineq, row)
-            push!(linkJineq, var.index.value)
-            push!(linkVineq, coeff)
-        end
-    end
-
-    #EQUALITY CONSTRAINTS
-    for (idx,link) in enumerate(linkeqconstraints)
-        row = idx
-        #Equality bounds
-        eqlink_lb[row] = link.set.value
-        eqlink_ub[row] = link.set.value
-
-        #Equality values
-        for (var,coeff) in link.func.terms
-            node = getnode(var)
-            local_data = _get_pips_data(node)
-            linkIeq = local_data.linkIeq
-            linkJeq = local_data.linkJeq
-            linkVeq = local_data.linkVeq
-            push!(linkIeq, row)              #the variable row
-            push!(linkJeq, var.index.value)  #the variable column
-            push!(linkVeq, coeff)            #the coefficient
-        end
-    end
-
-    #Add constraint data for each node to help with function evaluations
-
+    #Unpack needed data here
 
 
     #Initialize
@@ -388,12 +334,6 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
 		    return (local_data.n,local_data.m)
 		else
 		    if mode == :Values
-                # println("row_lb: ",row_lb)
-                # println("row_ub: ",row_ub)
-                # println(eqlink_lb)
-                # println(eqlink_ub)
-                # println(ineqlink_lb)
-                # println(ineqlink_ub)
                 original_copy([eqlink_lb; ineqlink_lb], row_lb)
                 original_copy([eqlink_ub; ineqlink_ub], row_ub)
             end
@@ -538,11 +478,11 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
                     Iineq=[node.ext[:Ijacineq];local_data.secondIineq .+ local_m_ineq]
                     Jineq=[node.ext[:Jjacineq];local_data.secondJineq]
                     Vineq=[local_values[jac_ineq_index];local_data.secondVineq]
-                    eqmat = sparseKeepZero(Ieq, Jeq, Veq, local_m_eq + local_data.num_eqconnect, local_data.n)
-                    ineqmat = sparseKeepZero(Iineq, Jineq, Vineq, local_m_ineq + local_data.num_ineqconnect, local_data.n)
+                    eqmat = _sparse_keep_zero(Ieq, Jeq, Veq, local_m_eq + local_data.num_eqconnect, local_data.n)
+                    ineqmat = _sparse_keep_zero(Iineq, Jineq, Vineq, local_m_ineq + local_data.num_ineqconnect, local_data.n)
                 else #evaluate
-                    eqmat = sparseKeepZero(local_m_eq .+ local_data.firstIeq, local_data.firstJeq, local_data.firstVeq, local_m_eq .+ local_data.num_eqconnect, first_stage_data.n)
-                    ineqmat = sparseKeepZero(local_m_ineq .+ local_data.firstIineq, local_data.firstJineq, local_data.firstVineq, local_m_ineq .+ local_data.num_ineqconnect, first_stage_data.n)
+                    eqmat = _sparse_keep_zero(local_m_eq .+ local_data.firstIeq, local_data.firstJeq, local_data.firstVeq, local_m_eq .+ local_data.num_eqconnect, first_stage_data.n)
+                    ineqmat = _sparse_keep_zero(local_m_ineq .+ local_data.firstIineq, local_data.firstJineq, local_data.firstVineq, local_m_ineq .+ local_data.num_ineqconnect, first_stage_data.n)
                 end
                 if length(eqmat.nzval) > 0
                     array_copy(eqmat.rowval,e_rowidx)
@@ -586,7 +526,6 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
     end
 
     function str_eval_h(rowid,colid,x0,x1,obj_factor,lambda,mode,rowidx,colptr,values)
-        #println("eval_h")
         node = model_list[colid+1]
         local_data = _get_pips_data(node)
         if mode == :Structure
@@ -594,14 +533,14 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
                 node_Hrows  = node.ext[:Hrows]
                 node_Hcols = node.ext[:Hcols]
                 node_val = ones(Float64,length(node_Hrows))
-                mat = sparseKeepZero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
+                mat = _sparse_keep_zero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
                 return length(mat.rowval)
             elseif colid == 0
                 node_Hrows  = node.ext[:Hrows]
                 node_Hcols = node.ext[:Hcols]
-                node_Hrows,node_Hcols = exchange(node_Hrows,node_Hcols)
+                node_Hrows,node_Hcols = _exchange(node_Hrows,node_Hcols)
                 node_val = ones(Float64,length(node_Hrows))
-                mat = sparseKeepZero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
+                mat = _sparse_keep_zero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
                 return length(mat.rowval)
             else
                 return 0
@@ -639,8 +578,8 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
                         local_sym_index +=1
                     end
                 end
-                node_Hrows,node_Hcols = exchange(node_Hrows,node_Hcols)
-                mat = sparseKeepZero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
+                node_Hrows,node_Hcols = _exchange(node_Hrows,node_Hcols)
+                mat = _sparse_keep_zero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
                 array_copy(mat.rowval,rowidx)
                 array_copy(mat.colptr,colptr)
                 original_copy(mat.nzval,values)
@@ -648,8 +587,8 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
                 node_Hrows  = node.ext[:Hrows]
                 node_Hcols = node.ext[:Hcols]
                 node_val = zeros(Float64,length(node_Hrows))
-                node_Hrows,node_Hcols = exchange(node_Hrows,node_Hcols)
-                mat = sparseKeepZero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
+                node_Hrows,node_Hcols = _exchange(node_Hrows,node_Hcols)
+                mat = _sparse_keep_zero(node_Hrows,node_Hcols,node_val, local_data.n, local_data.n)
                 array_copy(mat.rowval,rowidx)
                 array_copy(mat.colptr,colptr)
                 original_copy(mat.nzval,values)
@@ -728,13 +667,16 @@ function pipsnlp_solve(graph::OptiGraph) #Assume graph variables and constraints
             MPI.Bcast!(local_data.x_sol, local_data.n, coreid[1], comm)
         end
         #node.ext[:colVal] = local_data.x_sol
-        vars = MOI.get(node,MOI.ListOfVariableIndices())
-        primals = OrderedDict(zip(vars,local_data.x_sol))
-
-        id = graph.id
-        src = JuMP.backend(node)
-        src.primals[id] = primals
-        src.last_solution_id = id
+        #vars = MOI.get(node,MOI.ListOfVariableIndices())
+        vars = Plasmo.all_variables(node)
+        vals = local_data.x_sol
+        # primals = OrderedDict(zip(vars,local_data.x_sol))
+        Plasmo.set_node_primals(node,vars,vals)
+        #use API function
+        # id = graph.id
+        # src = JuMP.backend(node)
+        # src.primals[id] = primals
+        # src.last_solution_id = id
 
         #TODO set duals
